+++ clients/admin-ui/src/pages/dataset/index.tsx
@@ -74,10 +74,13 @@ const DataSets: NextPage = () => {
   } = useServerSidePagination();
 
   const [globalFilter, setGlobalFilter] = useState<string>();
-  const updateGlobalFilter = (searchTerm: string) => {
-    resetPageIndexToDefault();
-    setGlobalFilter(searchTerm);
-  };
+  const updateGlobalFilter = useCallback(
+    (searchTerm: string) => {
+      resetPageIndexToDefault();
+      setGlobalFilter(searchTerm);
+    },
+    [resetPageIndexToDefault, setGlobalFilter],
+  );
 
   const {
     data: datasetResponse,